requirements.txt
python-dotenv
requests
requests-oauthlib
setuptools
